requirements.txt
paramiko==3.4.0
PyYAML==6.0.1
requests==2.31.0